Fix precision issues with binance api

+++ priv/repo/migrations/20180305233120_create_trading_trades.exs
@@ -9,7 +9,7 @@ defmodule Pamela.Repo.Migrations.CreateTradingTrades do
       add(:coin, :string)
       add(:price, :string)
       add(:amount, :string)
-      add(:order_id, :integer)
+      add(:order_id, :bigint)
       add(:time, :utc_datetime)
 
       timestamps()
